#28714 Fix Unittest for l10nmgr, added optional ext_dependencies
git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/l10nmgr/branches/v4@84029 735d13b6-9817-0410-8766-e36946ffe9aa

+++ tests/mixed/class.tx_l10nmgr_mixed_completeWorkflow_testcase.php
@@ -69,6 +69,15 @@ class tx_l10nmgr_mixed_completeWorkflow_testcase extends tx_l10nmgr_tests_databa
 			'static_info_tables',
 			'templavoila',
 		);
+
+		// Read extension dependencies from extension configuration
+		$extConfigurationArray = unserialize($GLOBALS['TYPO3_CONF_VARS']['EXT']['extConf']['l10nmgr']);
+		if(isset($extConfigurationArray['ext_dependencies'])){
+			$dependencyArray = explode(',', $extConfigurationArray['ext_dependencies']);
+			$optional = array_merge($optional, $dependencyArray);
+			$optional = array_unique($optional);
+		}
+
 		foreach ($optional as $ext) {
 			if (t3lib_extMgm::isLoaded($ext)) {
 				$import[] = $ext;
@@ -617,13 +626,17 @@ class tx_l10nmgr_mixed_completeWorkflow_testcase extends tx_l10nmgr_tests_databa
 
 		if(is_array($files)){
 			foreach(array_reverse($files) as $file){
-				unlink($file);
+				if(file_exists($file)){
+					unlink($file);
+				}
 			}
 		}
 
 		if(is_array($directorys)){
 			foreach(array_reverse($directorys) as $directory){
-				rmdir($directory);
+				if(file_exists($directory)){
+					rmdir($directory);
+				}
 			}
 		}
 	}
